Implemented activity for adding a new crash when is detected by the accellerometer via notification

+++ app/src/main/java/com/example/crashcontrol/data/database/Entities.kt
@@ -31,5 +31,5 @@ data class Crash(
     val height: Double,
 
     @ColumnInfo(name = "impact_accelleration")
-    val impactAccelleration: Double
+    val impactAccelleration: Float
 )
